\name{viz.locations}
\alias{viz.locations}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Visualize location data
}
\description{
Using a data frame of device location data, plot as a heat map.
}
\usage{
viz.locations(location.df, map = "world", region = ".") 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{location.df}{
A location data frame, as generated by get.mylocation
}
  \item{map}{
(character) Map name 
}
  \item{region}{
(character) Region name
}
}
\details{
For more information on the documentation for the map package, http://cran.r-project.org/web/packages/maps/maps.pdf
}
\value{
Plots a map to the current device
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Drew Conway
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
library(stalkR)
my.locs<-get.mylocations("mike", "Mike Dewar's iPhone")
viz.locations(my.locs, "state", "new york")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{aplot}
